@@ -1,37 +1,37 @@
 {
   "name": "svelte-app",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
-    "build": "rollup -c",
-    "dev": "rollup -c -w",
-    "start": "sirv public --no-clear"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "lint": "prettier --check . && eslint . --fix",
+    "format": "prettier --write ."
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^17.0.0",
-    "@rollup/plugin-image": "^2.1.1",
-    "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^11.0.0",
-    "prettier": "^2.4.1",
-    "prettier-plugin-svelte": "^2.4.0",
-    "rollup": "^2.3.4",
-    "rollup-plugin-css-only": "^3.1.0",
-    "rollup-plugin-inline-svg": "^3.0.3",
-    "rollup-plugin-livereload": "^2.0.0",
-    "rollup-plugin-scss": "^3.0.0",
-    "rollup-plugin-svelte": "^7.0.0",
-    "rollup-plugin-svelte-svg": "^1.0.0-beta.6",
-    "rollup-plugin-svg": "^2.0.0",
-    "rollup-plugin-terser": "^7.0.0",
-    "svelte": "^3.0.0"
+    "@fullhuman/postcss-purgecss": "^5.0.0",
+    "@iconify/json": "^2.2.30",
+    "@sveltejs/adapter-netlify": "^1.0.5",
+    "@sveltejs/vite-plugin-svelte": "^2.0.2",
+    "eslint": "^8.35.0",
+    "eslint-config-prettier": "^8.7.0",
+    "postcss": "^8.4.21",
+    "prettier": "2.8.4",
+    "prettier-plugin-svelte": "^2.9.0",
+    "sass": "^1.57.1",
+    "svelte": "^3.55.1",
+    "unplugin-icons": "^0.15.3",
+    "vite": "^4.0.4",
+    "vite-imagetools": "^4.0.18"
   },
   "dependencies": {
-    "@ibm/plex": "^6.0.2",
-    "@popperjs/core": "^2.11.5",
-    "bootstrap": "^5.1.3",
-    "gsap": "^3.10.4",
-    "node-sass": "^7.0.1",
-    "sirv-cli": "^1.0.0",
-    "svelte-preprocess": "^4.10.7"
+    "@fontsource/ibm-plex-sans": "^4.5.13",
+    "@popperjs/core": "^2.11.6",
+    "@sveltejs/kit": "^1.8.5",
+    "animate.css": "^4.1.1",
+    "bootstrap": "^5.2.3",
+    "svelte-preprocess": "^5.0.0"
   }
 }
